requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==2.1.3
Flask-Cors==5.0.0
fonttools==4.55.3
frozenlist==1.5.0
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
nltk==3.9.1
numpy==2.2.1
openai==0.27.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
Werkzeug==2.0.3
yarl==1.18.3
